minor project structure
